requirements.txt
audio_recorder_streamlit
chainlit
deepgram-sdk
easyocr
fastembed
fastembed
langchain
langchain-community
langchain-groq
llama-parse
numpy
opencv-python
pandas
pillow
python-dotenv
qdrant_client
requests
streamlit
streamlit_float
unstructured[md]